+++ example/lib/routes.dart
@@ -4,13 +4,13 @@ import 'package:flutter/material.dart';
 import 'package:qlevar_router/qlevar_router.dart';
 
 import 'helpers/text_page.dart';
+import 'routes/overlays.dart';
 import 'routes/parent.dart';
 import 'screens/add_remove_routes.dart';
 import 'screens/declarative_page.dart';
 import 'screens/examples_page.dart';
 import 'screens/home_page.dart';
 import 'screens/nested_route.dart';
-import 'screens/overlays_page.dart';
 
 class AppRoutes {
   static const nested = 'Nested';
@@ -27,11 +27,11 @@ class AppRoutes {
 
   List<QRoute> routes() => [
         QRoute(path: '/', builder: () => HomePage()),
-        QRoute(path: '/overlays', builder: () => OverlaysPage()),
         QRoute(path: '/examples', builder: () => ExamplesPage()),
         QRoute.declarative(
             path: '/declarative',
             declarativeBuilder: (k) => DeclarativePage(k)),
+        OverlaysRoutes().route(),
         ParentRoutes().route(),
         QRoute(
             path: '/:id',