+++ lib/qlevar_router.dart
@@ -2,6 +2,7 @@ import 'src/qr.dart';
 
 export 'src/overlays/qdialog.dart';
 export 'src/overlays/qnotification.dart';
+export 'src/overlays/qpanel.dart';
 export 'src/pages/qpages.dart';
 export 'src/routers/qdeclarative._router.dart';
 export 'src/routers/qrouter.dart';