+++ lib/src/overlays/qdialog.dart
@@ -5,7 +5,7 @@ import '../routes/qdialog_route.dart';
 import 'qoverlay.dart';
 
 // Create a dialog to use with [QR]
-class QDialog with QOverlay {
+class QDialog {
   /// It gives the pop action to close the dialog and takes the widget to show
   final Widget Function(void Function<T>([T]) pop) widget;
   final bool barrierDismissible = true;
@@ -50,45 +50,51 @@ class QDialog with QOverlay {
             ));
   }
 
+  // @override
+  // Future<T?> show<T>(
+  //     {String? name, NavigatorState? state, BuildContext? context}) async {
+  //   if (state == null || context == null) {
+  //     if (name == null) {
+  //       return await QR.rootNavigator.show(this);
+  //     } else {
+  //       return await QR.navigatorOf(name).show(this);
+  //     }
+  //   }
+
+  //   QR.isShowingDialog = true;
+  //   final result = await Future.delayed(
+  //       Duration.zero,
+  //       () => state.push<T>(QDialogRoute(
+  //             pageBuilder: (buildContext, animation, secondaryAnimation) =>
+  //                 useSafeArea
+  //                     ? SafeArea(child: widget(state.pop))
+  //                     : widget(state.pop),
+  //             barrierDismissible: barrierDismissible,
+  //             barrierLabel:
+  //                 MaterialLocalizations.of(context).modalBarrierDismissLabel,
+  //             barrierColor: barrierColor ?? Colors.black54,
+  //             transitionDuration:
+  //                 transitionDuration ?? const Duration(milliseconds: 300),
+  //             transitionBuilder: transitionBuilder ??
+  //                 (context, animation, secondaryAnimation, child) {
+  //                   return FadeTransition(
+  //                     opacity: CurvedAnimation(
+  //                       parent: animation,
+  //                       curve: transitionCurve ?? Curves.easeOutQuad,
+  //                     ),
+  //                     child: child,
+  //                   );
+  //                 },
+  //             settings: routeSettings,
+  //           )));
+  //   QR.isShowingDialog = false;
+
+  //   return result;
+  // }
+
   @override
-  Future<T?> show<T>(
-      {String? name, NavigatorState? state, BuildContext? context}) async {
-    if (state == null || context == null) {
-      if (name == null) {
-        return await QR.rootNavigator.show(this);
-      } else {
-        return await QR.navigatorOf(name).show(this);
-      }
-    }
-
-    QR.isShowingDialog = true;
-    final result = await Future.delayed(
-        Duration.zero,
-        () => state.push<T>(QDialogRoute(
-              pageBuilder: (buildContext, animation, secondaryAnimation) =>
-                  useSafeArea
-                      ? SafeArea(child: widget(state.pop))
-                      : widget(state.pop),
-              barrierDismissible: barrierDismissible,
-              barrierLabel:
-                  MaterialLocalizations.of(context).modalBarrierDismissLabel,
-              barrierColor: barrierColor ?? Colors.black54,
-              transitionDuration:
-                  transitionDuration ?? const Duration(milliseconds: 300),
-              transitionBuilder: transitionBuilder ??
-                  (context, animation, secondaryAnimation, child) {
-                    return FadeTransition(
-                      opacity: CurvedAnimation(
-                        parent: animation,
-                        curve: transitionCurve ?? Curves.easeOutQuad,
-                      ),
-                      child: child,
-                    );
-                  },
-              settings: routeSettings,
-            )));
-    QR.isShowingDialog = false;
-
-    return result;
+  List<OverlayEntry> buildEntries<T>(BuildContext context) {
+    // TODO: implement build
+    throw UnimplementedError();
   }
 }